requirements.txt
nltk==3.8.1
pdf2image==1.16.3
google.cloud
tarfile
TexSoup
